# [Core] Use array to speedup padding