@casper-hansen 2024-07-25
Nice to see an 18% speedup from this optimization. Is it mainly for small models?