@peng1999 2024-07-25
> Is it mainly for small models?

Yes. This PR is for small models and large batch sizes. The `from_sampling_metadata` function, optimized by this PR, primarily runs on the CPU and is independent of logistics. Therefore, it can overlap with the GPU work of model inference. It will only be on the critical path if its execution time exceeds that of model inference, which occurs with smaller models.