@daquexian 2024-07-25
Great PR! Would you mind sharing what tool you used to get this image, is it nsight system?

![image](https://github.com/user-attachments/assets/d35f7511-da3c-41fd-b0fe-30034136d468)