@peng1999 2024-07-26
> > Great PR! Would you mind sharing what tool you used to get this image, is it nsight system?
> 
> Yes. The blue spans are recorded using NVTX.